policy: move common code into policy_common.go
No code changes, just moving things around.
